+++ ui/src/services/rpb.rs
@@ -1,5 +1,6 @@
 use anyhow::Result;
-use core::GetWeightsQuery;
+use core::{GetWeightsQuery, SearchQuery};
+use serde::{Deserialize, Serialize};
 use yew::{services::fetch::FetchTask, Callback};
 
 #[derive(PartialEq)]
@@ -9,6 +10,15 @@ pub struct Service {
 
 impl super::Service for Service {}
 
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct TickerInfo {
+  pub symbol: String,
+  pub name: String,
+  pub exch_disp: String,
+  pub type_disp: String,
+}
+
 impl Service {
   pub fn get_weigths(
     &self,
@@ -16,6 +26,14 @@ impl Service {
     callback: Callback<Result<Vec<f64>>>,
   ) -> FetchTask {
     super::Service::get(self, &self.prepend_base("weights"), Some(&query), callback)
+  }
+
+  pub fn get_search(
+    &self,
+    query: SearchQuery,
+    callback: Callback<Result<Vec<TickerInfo>>>,
+  ) -> FetchTask {
+    super::Service::get(self, &self.prepend_base("search"), Some(&query), callback)
   }
 
   // TODO this can be taken away with a new PathBuilder abstraction
